fix two full_like cases

+++ framework/api/paddlebase/test_full_like.py
@@ -177,10 +177,10 @@ def test_full_like13():
     """
     fill_value is inf
     """
-    x = np.array([2, 3])
+    x = np.array([2.0, 3.0])
     fill_value = float("inf")
     res = np.full_like(x, fill_value)
-    obj.run(res=res, x=x, fill_value=fill_value, dtype="int64")
+    obj.run(res=res, x=x, fill_value=fill_value)
 
 
 @pytest.mark.api_base_full_like_parameters
@@ -188,7 +188,7 @@ def test_full_like14():
     """
     fill_value is -inf
     """
-    x = np.array([2, 3])
+    x = np.array([2.0, 3.0])
     fill_value = float("-inf")
     res = np.full_like(x, fill_value)
-    obj.run(res=res, x=x, fill_value=fill_value, dtype="int64")
+    obj.run(res=res, x=x, fill_value=fill_value)
